requirements.txt
torch
numpy
Pillow
transformers>=4.30.1
sentence-transformers
decord
ffmpeg-python
scenedetect[opencv]
opencv-python
segment-anything
ultralytics
pytesseract
open-clip-torch
shapely
pycocotools
soundfile
groq
faiss-cpu
